+++ src/constants.js
@@ -2,9 +2,9 @@ import { ACTOR_EVENT_NAMES } from '@apify/consts';
 
 /**
  * The default user agent used by `Apify.launchPuppeteer`.
- * Last updated on 2020-05-22.
+ * Last updated on 2022-05-05.
  */
-export const DEFAULT_USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/87.0.4280.67 Safari/537.36';
+export const DEFAULT_USER_AGENT = 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/101.0.4951.54 Safari/537.36';
 
 /**
  * Exit codes for the actor process.