+++ test/crawlers/playwright_crawler.test.js
@@ -70,6 +70,8 @@ describe('PlaywrightCrawler', () => {
                 expect(await response.status()).toBe(200);
                 request.userData.title = await page.title();
                 processed.push(request);
+                expect(response.request().headers()['user-agent']).not.toMatch(/headless/i);
+                await expect(page.evaluate(() => window.navigator.webdriver)).resolves.toBeFalsy();
             };
 
             const playwrightCrawler = new Apify.PlaywrightCrawler({